Replace map with recordable time key type from long to int

+++ api/src/main/java/mc/replay/api/recording/RecordingSession.java
@@ -16,7 +16,7 @@ public interface RecordingSession {
 
     long getStartTime();
 
-    @NotNull NavigableMap<Long, List<CachedRecordable>> getRecordables();
+    @NotNull NavigableMap<Integer, List<CachedRecordable>> getRecordables();
 
     @NotNull Recording stopRecording();
 }
